Restructured, and created blank pages to fill out. Also modularized the Hero component from the LandingPage.js as a separate component.
