+++ client/src/components/header/index.js
@@ -1,8 +1,23 @@
 import React from 'react';
-import NavMain from './nav-main'
+import './navMain.css'
+import { Link } from "react-router-dom";
 
 export default function Header() {
+    
     return (
-        < NavMain />
+        <nav>
+            <div>
+                <button className='logo' onClick={() => alert("this is our home")}>[ &nbsp;]</button>
+            </div>
+            <div>
+                <button  onClick={() => alert("This is our website")}>Learn</button>
+            </div>
+            <div>
+                <button onClick={() => alert("Stay safe")}>Safety</button>
+            </div>
+            <div>
+                <button onClick={() => alert("Need help? Me too!!")}>Support</button>
+            </div>
+        </nav>
     )
 }
