+++ client/src/components/hero/index.js
@@ -1,0 +1,20 @@
+import React from 'react'
+
+function index() {
+    const styles = {
+        sloganContainer: {
+            borderLeft: "solid 2px #7B7B7B",
+            padding: "0 1em",
+            margin: "10px 2em"
+        },
+        }
+      
+
+  return (
+    <div className="sloganContainer" style={styles.sloganContainer}>
+        <h2>&#123; #COMMIT TO YOUR RIGHT DEVELOPER &#125;</h2>
+    </div>
+  )
+}
+
+export default index
